Update AvgResponseTime from float to int

+++ src/Models/User.cs
@@ -66,7 +66,7 @@ namespace StreamChat.Models
         public Language? Language { get; set; }
         public List<string> Teams { get; set; }
         public List<string> LatestHiddenChannels { get; set; }
-        public float? AvgResponseTime { get; set; }
+        public int? AvgResponseTime { get; set; }
     }
 
     public class AddMemberOptions
